requirements.txt
apache-airflow==2.7.3
psycopg2-binary
python-dotenv
pandas
great-expectations
dbt-core
dbt-postgres
Flask-Session==0.4.0